Refactor test dependencies to centralize cluster testing under server-cluster
