+++ src/app/clusters/electrical-power-measurement-server/tests/BUILD.gn
@@ -29,8 +29,8 @@ chip_test_suite("tests") {
 
   public_deps = [
     "${chip_root}/src/app/clusters/electrical-power-measurement-server",
-    "${chip_root}/src/app/clusters/testing",
     "${chip_root}/src/app/data-model-provider/tests:encode-decode",
+    "${chip_root}/src/app/server-cluster/testing",
     "${chip_root}/src/app/server-cluster/testing:testing",
     "${chip_root}/src/lib/support",
   ]
@@ -55,8 +55,8 @@ chip_test_suite("tests-backwards-compatibility") {
 
   public_deps = [
     "${chip_root}/src/app/clusters/electrical-power-measurement-server",
-    "${chip_root}/src/app/clusters/testing",
     "${chip_root}/src/app/data-model-provider/tests:encode-decode",
+    "${chip_root}/src/app/server-cluster/testing",
     "${chip_root}/src/app/server-cluster/testing:testing",
     "${chip_root}/src/data-model-providers/codegen/tests:mock_model",
     "${chip_root}/src/lib/support",
